Move api to `Nelmio_v3`
