+++ src/Resources/config/routing/api_nelmio3/group.xml
@@ -1,18 +1,18 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <routes xmlns="http://symfony.com/schema/routing" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/routing https://symfony.com/schema/routing/routing-1.0.xsd">
-    <route id="delete_group" path="/groups/{id}.{_format}" methods="DELETE" controller="Sonata\UserBundle\Controller\Api\GroupController::deleteGroupAction" format="json">
+    <route id="delete_group" path="/groups/{id}.{_format}" methods="DELETE" controller="Sonata\UserBundle\Controller\Api\Nelmio_v3\GroupController::deleteGroupAction" format="json">
         <requirement key="_format">json|xml|html</requirement>
     </route>
-    <route id="get_group" path="/groups/{id}.{_format}" methods="GET" controller="Sonata\UserBundle\Controller\Api\GroupController::getGroupAction" format="json">
+    <route id="get_group" path="/groups/{id}.{_format}" methods="GET" controller="Sonata\UserBundle\Controller\Api\Nelmio_v3\GroupController::getGroupAction" format="json">
         <requirement key="_format">json|xml|html</requirement>
     </route>
-    <route id="get_groups" path="/groups.{_format}" methods="GET" controller="Sonata\UserBundle\Controller\Api\GroupController::getGroupsAction" format="json">
+    <route id="get_groups" path="/groups.{_format}" methods="GET" controller="Sonata\UserBundle\Controller\Api\Nelmio_v3\GroupController::getGroupsAction" format="json">
         <requirement key="_format">json|xml|html</requirement>
     </route>
-    <route id="post_group" path="/groups.{_format}" methods="POST" controller="Sonata\UserBundle\Controller\Api\GroupController::postGroupAction" format="json">
+    <route id="post_group" path="/groups.{_format}" methods="POST" controller="Sonata\UserBundle\Controller\Api\Nelmio_v3\GroupController::postGroupAction" format="json">
         <requirement key="_format">json|xml|html</requirement>
     </route>
-    <route id="put_group" path="/groups/{id}.{_format}" methods="PUT" controller="Sonata\UserBundle\Controller\Api\GroupController::putGroupAction" format="json">
+    <route id="put_group" path="/groups/{id}.{_format}" methods="PUT" controller="Sonata\UserBundle\Controller\Api\Nelmio_v3\GroupController::putGroupAction" format="json">
         <requirement key="_format">json|xml|html</requirement>
     </route>
 </routes>
