+++ src/Resources/config/routing/api_nelmio3/user.xml
@@ -1,24 +1,24 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <routes xmlns="http://symfony.com/schema/routing" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/routing https://symfony.com/schema/routing/routing-1.0.xsd">
-    <route id="delete_user" path="/users/{id}.{_format}" methods="DELETE" controller="Sonata\UserBundle\Controller\Api\UserController::deleteUserAction" format="json">
+    <route id="delete_user" path="/users/{id}.{_format}" methods="DELETE" controller="Sonata\UserBundle\Controller\Api\Nelmio_v3\UserController::deleteUserAction" format="json">
         <requirement key="_format">json|xml|html</requirement>
     </route>
-    <route id="delete_user_group" path="/users/{userId}/groups/{groupId}.{_format}" methods="DELETE" controller="Sonata\UserBundle\Controller\Api\UserController::deleteUserGroupAction" format="json">
+    <route id="delete_user_group" path="/users/{userId}/groups/{groupId}.{_format}" methods="DELETE" controller="Sonata\UserBundle\Controller\Api\Nelmio_v3\UserController::deleteUserGroupAction" format="json">
         <requirement key="_format">json|xml|html</requirement>
     </route>
-    <route id="get_user" path="/users/{id}.{_format}" methods="GET" controller="Sonata\UserBundle\Controller\Api\UserController::getUserAction" format="json">
+    <route id="get_user" path="/users/{id}.{_format}" methods="GET" controller="Sonata\UserBundle\Controller\Api\Nelmio_v3\UserController::getUserAction" format="json">
         <requirement key="_format">json|xml|html</requirement>
     </route>
-    <route id="get_users" path="/users.{_format}" methods="GET" controller="Sonata\UserBundle\Controller\Api\UserController::getUsersAction" format="json">
+    <route id="get_users" path="/users.{_format}" methods="GET" controller="Sonata\UserBundle\Controller\Api\Nelmio_v3\UserController::getUsersAction" format="json">
         <requirement key="_format">json|xml|html</requirement>
     </route>
-    <route id="post_user" path="/users.{_format}" methods="POST" controller="Sonata\UserBundle\Controller\Api\UserController::postUserAction" format="json">
+    <route id="post_user" path="/users.{_format}" methods="POST" controller="Sonata\UserBundle\Controller\Api\Nelmio_v3\UserController::postUserAction" format="json">
         <requirement key="_format">json|xml|html</requirement>
     </route>
-    <route id="post_user_group" path="/users/{userId}/groups/{groupId}.{_format}" methods="POST" controller="Sonata\UserBundle\Controller\Api\UserController::postUserGroupAction" format="json">
+    <route id="post_user_group" path="/users/{userId}/groups/{groupId}.{_format}" methods="POST" controller="Sonata\UserBundle\Controller\Api\Nelmio_v3\UserController::postUserGroupAction" format="json">
         <requirement key="_format">json|xml|html</requirement>
     </route>
-    <route id="put_user" path="/users/{id}.{_format}" methods="PUT" controller="Sonata\UserBundle\Controller\Api\UserController::putUserAction" format="json">
+    <route id="put_user" path="/users/{id}.{_format}" methods="PUT" controller="Sonata\UserBundle\Controller\Api\Nelmio_v3\UserController::putUserAction" format="json">
         <requirement key="_format">json|xml|html</requirement>
     </route>
 </routes>
